📦 Moves statusbar to own component
