+++ app/src/lib/components/statusbar/ThemeButton.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+    import MdiWeatherSunny from '~icons/mdi/weather-sunny';
+	import MdiMoonAndStars from '~icons/mdi/moon-and-stars';
+</script>
+
+<label class="swap swap-rotate">
+    <input type="checkbox" value="light" class="theme-controller" />
+    <MdiWeatherSunny class="swap-off h-7 w-7" />
+    <MdiMoonAndStars class="swap-on h-7 w-7" />
+</label>
