+++ app/src/lib/components/statusbar/TopBar.svelte
